% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DutchWeather.R
\docType{package}
\name{DutchWeather}
\alias{DutchWeather}
\title{DutchWeather: A package for importing KKNMI weather data.}
\description{
The DutchWeather package provides functions to import weather statistics
from the KNMI website.
}
\section{KNMI daily weather data}{

\code{\link[DutchWeather]{get_daily_zipdata}} \cr
}

