% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Weerlive_API_data.R
\name{get_Weerlive_API_data}
\alias{get_Weerlive_API_data}
\title{Get real time data from the Weerlive.nl API}
\usage{
get_Weerlive_API_data(
  key,
  locatie = "Amsterdam",
  base = "https://weerlive.nl/api/json-data-10min.php?key=\%1.0f&locatie=\%s"
)
}
\arguments{
\item{key}{Int. Your API key}

\item{locatie}{Character. Your location written in dutch. Default: "Amsterdam"}

\item{base}{Character.}
}
\value{

}
\description{
Don't forget to get your API acces key by visiting http://weerlive.nl/.
}
